Refactor cookie-consent implementation
- Move Routes into separate component
- Move Page into separate component
- Remove react-cookie-consent dependency
- Implement cookie consent with React Context. They information about whether or not cookies are consented can now be accessed from any component using the React Context
- Implement sentry loading as component that will check cookie consent
- Use a hook for logging page changes if cookies are consented
