+++ package.json
@@ -19,7 +19,6 @@
     "i18next-xhr-backend": "^3.2.2",
     "postcss-cli": "^7.1.0",
     "react": "^16.13.0",
-    "react-cookie-consent": "^3.0.0",
     "react-dom": "^16.13.0",
     "react-firebase-hooks": "^2.1.1",
     "react-i18next": "^11.3.3",
